+++ src/text_to_sql_agent/services/__init__.py
@@ -6,6 +6,7 @@ from .schema_normalization import build_snapshot_id, normalize_raw_schema
 from .schema_document_builder import build_schema_documents
 from .schema_indexing import index_schema_embeddings
 from .query_result_export import export_query_result
+from .query_analytics import QueryAnalyticsResult, build_one_shot_chart
 
 __all__ = [
     "build_snapshot_id",
@@ -13,4 +14,6 @@ __all__ = [
     "index_schema_embeddings",
     "normalize_raw_schema",
     "export_query_result",
+    "QueryAnalyticsResult",
+    "build_one_shot_chart",
 ]